+++ packages/react-instantsearch/src/widgets/TrendingFacets.tsx
@@ -10,11 +10,7 @@ import type { UseTrendingFacetsProps } from 'react-instantsearch-core';
 
 type UiProps = Pick<
   TrendingFacetsUiComponentProps,
-  | 'items'
-  | 'itemComponent'
-  | 'headerComponent'
-  | 'emptyComponent'
-  | 'status'
+  'items' | 'itemComponent' | 'headerComponent' | 'emptyComponent' | 'status'
 >;
 
 export type TrendingFacetsProps = Omit<
@@ -37,7 +33,6 @@ export function TrendingFacets({
   limit,
   threshold,
   fallbackParameters,
-  queryParameters,
   escapeHTML,
   transformItems,
   itemComponent,
@@ -52,7 +47,6 @@ export function TrendingFacets({
       limit,
       threshold,
       fallbackParameters,
-      queryParameters,
       escapeHTML,
       transformItems,
     },